TD: adjust types for multiplayer

+++ TD/session.h
@@ -289,28 +289,28 @@ typedef struct {
 		struct {
 			HousesType House;						// player's House
 			unsigned int Color;					// player's color
-			unsigned long NameCRC;				// CRC of player's game's name
-			unsigned long MinVersion;			// game's min supported version
-			unsigned long MaxVersion;			// game's max supported version
+			uint32_t NameCRC;					// CRC of player's game's name
+			uint32_t MinVersion;				// game's min supported version
+			uint32_t MaxVersion;				// game's max supported version
 		} PlayerInfo;
 		struct {
 			unsigned char Scenario;				// Scenario #
 			unsigned int Credits;				// player's credits
-			unsigned int IsBases		: 1;		// 1 = bases are allowed
-			unsigned int IsTiberium	: 1;		// 1 = tiberium is allowed
-			unsigned int IsGoodies	: 1;		// 1 = goodies are allowed
-			unsigned int IsGhosties	: 1;		// 1 = ghosts are allowed
+			uint8_t IsBases		: 1;			// 1 = bases are allowed
+			uint8_t IsTiberium	: 1;			// 1 = tiberium is allowed
+			uint8_t IsGoodies	: 1;			// 1 = goodies are allowed
+			uint8_t IsGhosties	: 1;			// 1 = ghosts are allowed
 			unsigned char BuildLevel;			// buildable level
 			unsigned char UnitCount;			// max # units
 			int Seed;								// random number seed
 			SpecialClass Special;				// command-line options
 			unsigned int GameSpeed;				// Game Speed
-			unsigned long Version;				// version # common to all players
+			uint32_t Version;					// version # common to all players
 		} ScenarioInfo;
 		struct {
 			char Buf[COMPAT_MESSAGE_LENGTH];	// inter-user message
 			unsigned char Color;					// color of sender of message
-			unsigned long NameCRC;				// CRC of sender's Game Name
+			uint32_t NameCRC;					// CRC of sender's Game Name
 		} Message;
 		struct {
 			int OneWay;								// one-way response time
@@ -319,7 +319,7 @@ typedef struct {
 			int Why;									// why were we rejected from the game?
 		} Reject;
 		struct {
-			unsigned long ID;						// unique ID for this chat node
+			uint32_t ID;							// unique ID for this chat node
 			unsigned char Color;					// my color
 		} Chat;
 	};
